python-research-service/requirements.txt
gpt-researcher==0.10.0
fastapi==0.115.6
uvicorn==0.34.0
python-dotenv==1.0.1
pydantic==2.10.4
httpx==0.28.0